chore: Move most of keyboard.rs to winit-core
