+++ winit-core/src/lib.rs
@@ -12,3 +12,4 @@ compile_error! { "no-std and no-alloc usage are not yet supported" }
 
 pub mod dpi;
 pub mod error;
+pub mod keyboard;
