+++ winit/Cargo.toml
@@ -61,7 +61,7 @@ rwh_06 = ["dep:rwh_06", "ndk/rwh_06"]
 cfg_aliases = "0.2.0"
 
 [dependencies]
-bitflags = "2"
+bitflags.workspace = true
 cursor-icon = "1.1.0"
 log = "0.4"
 mint = { version = "0.5.6", optional = true }
@@ -70,7 +70,7 @@ rwh_04 = { package = "raw-window-handle", version = "0.4", optional = true }
 rwh_05 = { package = "raw-window-handle", version = "0.5.2", features = ["std"], optional = true }
 rwh_06 = { package = "raw-window-handle", version = "0.6", features = ["std"], optional = true }
 serde = { workspace = true, optional = true }
-smol_str = "0.2.0"
+smol_str.workspace = true
 winit-core.workspace = true
 
 [dev-dependencies]
